+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Fixed
 - Fixed crashes in the `pcs host auth` command ([rhbz#1676957])
+- Fixed id conflict with current bundle configuration in
+  `pcs resource bundle reset` ([rhbz#1657166])
 
 ### Changed
 - Command `pcs resource bundle reset` no longer accepts the container type
   ([rhbz#1657166]
-
 [rhbz#1657166]: https://bugzilla.redhat.com/show_bug.cgi?id=1657166
 [rhbz#1676957]: https://bugzilla.redhat.com/show_bug.cgi?id=1676957